requirements.txt
Flask>=2.3.0
requests>=2.31.0
mysql-connector-python>=8.0.0
reportlab>=4.0.0
qrcode>=7.4.0
Pillow>=8.0.0